requirements.txt
autopep8==1.5.4
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Flask==1.1.2
graphviz==0.15
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
neat-python==0.92
numpy==1.19.4
Pillow==8.0.1
pycodestyle==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
